feat: main page 구조변경
