+++ src/components/main/FormSection.js
@@ -1,0 +1,10 @@
+const FormSection = ({ title, children }) => {
+  return (
+    <div>
+      <h3>{title}</h3>
+      {children}
+    </div>
+  );
+};
+
+export default FormSection;
